% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Causal Coherence.R
\name{CalculateCausalPhase}
\alias{CalculateCausalPhase}
\title{Calculate phase difference from causal coherence}
\usage{
CalculateCausalPhase(ccoh)
}
\arguments{
\item{ccoh}{computed causal coherence}
}
\value{
the computed phase difference
}
\description{
Calculate phase difference from causal coherence
}
\examples{
#ADD EXAMPLE
}
\author{
Alvaro Chao-Ecija

Marc Stefan Dawid-Milner
}
