% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coherence.R
\name{CalculateCoherence}
\alias{CalculateCoherence}
\title{Calculate spectral coherence}
\usage{
CalculateCoherence(SM, index1, index2)
}
\arguments{
\item{SM}{the computed frequency domain model from the system to be analyzed}

\item{index1}{numeric index to specify a variable from the model}

\item{index2}{numeric index to specify a variable from the model}
}
\value{
The spectral coherence between the two variables
}
\description{
Calculates the spectral coherence between two variables
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)

coherence <- CalculateCoherence(freq_model, 1, 2)

}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
