% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CardioRVAR-package.R
\docType{package}
\name{CardioRVAR-package}
\alias{CardioRVAR}
\alias{CardioRVAR-package}
\title{CardioRVAR: Spectral Analysis of Open and Closed-Loop Cardiovascular Interactions}
\description{
This is an R package designed for modelling cardiovascular open- and closed-loop interactions. It is advisable that this package is used with the separate ShinyApp interface, CarioRVARapp, for a better user-friendly experience.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CIMES-USNA-UMA/CardioRVAR}
  \item Report bugs at \url{https://github.com/CIMES-USNA-UMA/CardioRVAR/issues}
}

}
\keyword{internal}
