% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation_for_data.R
\docType{data}
\name{Cardiovascular}
\alias{Cardiovascular}
\title{Raw Cardiovascular Data}
\format{
A list with 5 variables:
\describe{
  \item{Time}{Time values}
  \item{RR}{RR intervals}
  \item{SBP}{Beat-to-beat systolic blood pressure}
  \item{DBP}{Beat-to-beat dyastolic blood pressure}
  \item{HR}{Heart rate values}
}
}
\usage{
data(Cardiovascular)
}
\description{
Data object containing systolic and diastolic blood pressure values, as well as RR intervals
and heart rate values.
}
\examples{
data(Cardiovascular)

}
\keyword{dataset}
