% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{CheckStationarity}
\alias{CheckStationarity}
\title{Check stationarity}
\usage{
CheckStationarity(
  x,
  alpha = 0.05,
  lags = NULL,
  warnings = TRUE,
  verbose = FALSE,
  correction = "bonferroni"
)
}
\arguments{
\item{x}{a matrix with the data to be checked. Number of columns should be equal
to the number of variables.}

\item{alpha}{significance level for testing. Default is 0.05}

\item{lags}{specify a certain level of lags for the test. Default is NULL}

\item{warnings}{boolean, show warnings regarding the test results. Default is TRUE}

\item{verbose}{boolean, show results from the test in captions. Default is FALSE}

\item{correction}{choose a p-value correction method for multiple hypotheses. Default
is bonferroni. For other methods, please check \link[stats]{p.adjust.methods} from
package \href{https://CRAN.R-project.org/package=stats}{stats}}
}
\value{
A boolean variable indicating if the multivariate time series is stationary or not,
        or a caption indicating so, depending on whether the arguments verbose and warning are TRUE or
        FALSE
}
\description{
Check stationarity of the data
}
\examples{
data(DetrendedData)

CheckStationarity(DetrendedData)
CheckStationarity(DetrendedData, verbose = TRUE)
}
\references{
Seth AK. A MATLAB toolbox for Granger causal connectivity analysis. J 
Neurosci Methods. 2010;186(2):262-73.

Seth AK. Granger Causal Conectivity Analysis: A MATLAB Toolbox. 2011.

Ding M, Bressler S, Yang W, Liang H. Short-window spectral analysis of cortical 
event-related potentials by adaptative multivariate autoregressive modelling: data
preprocessing, model validation, and variability assessment. Biol Cybern. 2000;83:35-45
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
