% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detrend.R
\name{DetrendByCutoff}
\alias{DetrendByCutoff}
\title{Detrend variables}
\usage{
DetrendByCutoff(x, cutoff = 0.04, f = 4, wv = "d16", max_f = 0.4)
}
\arguments{
\item{x}{A vector of observations}

\item{cutoff}{A frequency used as limit for the detrending. Default is 0.04 Hz}

\item{f}{Sample rate of the vector of observations. Default is 4 Hz}

\item{wv}{Wavelet to be passed into the function \link[waveslim]{modwt} of package
\href{https://CRAN.R-project.org/package=waveslim}{waveslim}. Default is
d16}

\item{max_f}{Maximum frequency to be accounted for in the detrending. default is 0.4 Hz}
}
\value{
The detrended vector of observations
}
\description{
Detrends a vector of observations using the discrete wavelet transform
}
\details{
This function uses a wavelet detrending algorithm to select the aproppriate componets of a signal
and discard low frequency components. This function requires package \href{https://CRAN.R-project.org/package=waveslim}{waveslim}:
for more information regarding this package, please check the references section.
}
\examples{
data(Cardiovascular)
int_data <- InterpolateData(Cardiovascular)
RR <- DetrendByCutoff(int_data$RR) # The RR series is detrended
SBP <- DetrendByCutoff(int_data$SBP) # The SBP series is detrended
}
\references{
Brandon Whitcher (2020). waveslim: Basic Wavelet Routines for One-, Two-, and Three-Dimensional Signal
Processing. R package version 1.8.2. https://CRAN.R-project.org/package=waveslim

Li L, Liu C, Li K, Liu C. Comparison of Detrending Methods in Spectral Analysis of Heart Rate Variability.
Res J Appl Sci Eng Technol. 2011;3(9):1014-21
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
