% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation_for_data.R
\docType{data}
\name{DetrendedData}
\alias{DetrendedData}
\title{Detrended Cardiovascular Data}
\format{
A matrix with 2 columns:
\describe{
  \item{SBP}{Detrended systolic blood pressure series}
  \item{RR}{Detrended RR intervals series}
}
}
\usage{
data(DetrendedData)
}
\description{
Detrended RR intervals and systolic blood pressure from \link[CardioRVAR]{Cardiovascular}
}
\examples{
data(DetrendedData)

}
\keyword{dataset}
