% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_Stats.R
\name{DiagnoseStability}
\alias{DiagnoseStability}
\title{Check stability of a VAR model}
\usage{
DiagnoseStability(
  model,
  do.plot = FALSE,
  verbose = FALSE,
  col = "blue",
  col2 = "red",
  verbose.method = c("c", "p")
)
}
\arguments{
\item{model}{a VAR model to be checked}

\item{do.plot}{boolean, plot a unitary circle showing the inverse roots of the
model. Default is FALSE}

\item{verbose}{boolean, show results from the test in captions. Default is FALSE}

\item{col}{color for the roots inside the unit circle. Default is blue}

\item{col2}{color for the roots outside the unit circle. Default is red}

\item{verbose.method}{method for the verbose option, either "c" (cat) or
"p" (print). Default is c.}
}
\value{
A boolean variable indicating if the multivariate time series is stationary or not.
}
\description{
Check that the estimated VAR model is stable
}
\examples{
data(DetrendedData)

model <- EstimateVAR(DetrendedData)
DiagnoseStability(model, do.plot = TRUE)
}
\references{
Seth AK. A MATLAB toolbox for Granger causal connectivity analysis. J 
Neurosci Methods. 2010;186(2):262-73.

Seth AK. Granger Causal Conectivity Analysis: A MATLAB Toolbox. 2011.

Ding M, Bressler S, Yang W, Liang H. Short-window spectral analysis of cortical 
event-related potentials by adaptative multivariate autoregressive modelling: data
preprocessing, model validation, and variability assessment. Biol Cybern. 2000;83:35-45
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner

#' @details 
Several functions are available in this package to validate
properties of the evaluated time series and VAR models. In this function, 
stability of the VAR model is evaluated according to the criteria
established in Ding (2000).
}
