% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expected_Values_Freq_Domain.R
\name{GetEstimateAtMaxCoh}
\alias{GetEstimateAtMaxCoh}
\title{Get estimates at maximum coherence}
\usage{
GetEstimateAtMaxCoh(SM, VLF = 0.04, LF = 0.15, HF = 0.4, str = TRUE, coherence)
}
\arguments{
\item{SM}{the computed frequency domain model from the system to be analyzed}

\item{VLF}{maximum limit for the VLF band. Default is 0.04}

\item{LF}{maximum limit for the LF band. Default is 0.15}

\item{HF}{maximum limit for the HF band. Default is 0.4}

\item{str}{boolean, if TRUE the output will be the structure of the returned R object.
Default is FALSE}

\item{coherence}{the spectral coherence between the analyzed variables}
}
\value{
A list containing the estimates at the HF and LF bands; or
        the structure of this list.
}
\description{
Use the coherence between two variables to get estimates of the frequency model
at each band at maximum coherence levels
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)
coherence <- CalculateCoherence(freq_model, 1, 2)
GetEstimateAtMaxCoh(freq_model, coherence = coherence)
Estimates <- GetEstimateAtMaxCoh(freq_model, coherence = coherence,
str = FALSE)
Estimates

}
\references{
Bari V, Vaini E, De Maria B, Cairo B, Pistuddi V, Ranucci M, Porta A. 
Comparison of Different Strategies to Assess Cardiac Baroreflex Sensitivity 
Based on Transfer Function Technique in Patients Undergoing General Anesthesia. 
Annu Int Conf IEEE Eng Med Biol Soc 2018. 2018;2780–3, doi:10.1109/EMBC.2018.8512782.
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
