% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Freq Domain Analysis.R
\name{GetExpectedValues}
\alias{GetExpectedValues}
\title{Compute expected values}
\usage{
GetExpectedValues(
  SM,
  VLF = 0.04,
  LF = 0.15,
  HF = 0.4,
  str = TRUE,
  weight = TRUE,
  use.coh = TRUE,
  thr = 0.5,
  coherence
)
}
\arguments{
\item{SM}{the computed frequency domain model from the system to be analyzed}

\item{VLF}{maximum limit for the VLF band. Default is 0.04}

\item{LF}{maximum limit for the LF band. Default is 0.15}

\item{HF}{maximum limit for the HF band. Default is 0.4}

\item{str}{boolean, if TRUE the output will be the structure of the returned R object.
Default is FALSE}

\item{weight}{applies gaussian-weighting function to the estimates before
calculating the expected values. Default is TRUE}

\item{use.coh}{boolean, if TRUE the coherence and coherence threshold will be
used for the computation of expected values. Default is TRUE}

\item{thr}{a coherence threshold to ensure the validity of the estimates. Default is 0.5}

\item{coherence}{the spectral coherence between the analyzed variables}
}
\value{
A list containing the expected values computed at the HF and LF bands; or
        the structure of this list.
}
\description{
Compute expected values from a frequency domain model of a particular system
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)

GetExpectedVals(freq_model, use.coh = FALSE)

ExpectedVals <- GetExpectedVals(freq_model, use.coh = FALSE, str = FALSE)

# If the coherence is to be used:
coherence <- CalculateCoherence(freq_model, 1, 2)

ExpectedVals <- GetExpectedVals(freq_model, weight = FALSE, coherence = coherence,
str = FALSE)

}
\references{
McLoone V, Ringwood JV. A system identification approach to baroreflex sensitivity
estimation
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
