% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_Modelling.R
\name{GetFreqs}
\alias{GetFreqs}
\title{Get vector of frequencies}
\usage{
GetFreqs(type, model, len, dt)
}
\arguments{
\item{type}{Method to get the vector of frequencies}

\item{model}{A time domain VAR model}

\item{len}{Length of the vector of frequencies}

\item{dt}{Inverse of sample rate}
}
\value{
Vector of frequency values
}
\description{
Get vector of frequencies
}
\details{
This function allows to obtain vectors of frequency values by either creating a sequence
or by spectral estimation. This second option was introduced to make the results compatible (and thus
reproducible) with the ones obtained from other packages such as package grangers (for more information
regarding this package, please check the reference section).
}
\examples{
#ADD EXAMPLE
}
\references{
Matteo Farne' and Angela Montanari (2019). grangers: Inference on Granger-Causality in the Frequency
Domain. R package version 0.1.0. https://CRAN.R-project.org/package=grangers
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
\keyword{internal}
