% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametric_Modelling.R
\name{GetMatrixAfromB}
\alias{GetMatrixAfromB}
\title{Calculate transfer function B from transfer function A}
\usage{
GetMatrixAfromB(B)
}
\arguments{
\item{B}{Transfer function B}
}
\value{
Transfer function A
}
\description{
Calculates transfer function B from a previously calculated transfer function A
}
\references{
Barbieri R, Parati G, Saul JP. Closed- versus Open-Loop Assessment of Heart Rate
Baroreflex. IEEE Eng Med Biol Mag. 2001;20(2):33-42

Korhonen I, Mainardi L, Baselli G, Bianchi A, Loula P, 
Carrault. Linear multivariate models for physiological signal analysis: applications. 
Comput Methods Programs Biomed. 1996;51(1-2):121-30

Faes L, Nollo G. Multivariate Frequency Domain Analysis of Causal Interactions
in Physiological Time Series. In Biomedical Engineering, Trends in Electronics,
Communications and Software. 2011

Hytti H, Takalo R, Ihalainen H. Tutorial on Multivariate Autoregressive Modelling.
J Clin Monit Comput. 2006;20(1):101-8
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
\keyword{internal}
