% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_Modelling.R
\name{GetMatrixBfromCoefs}
\alias{GetMatrixBfromCoefs}
\title{Calculate transfer function B from VAR coefficients}
\usage{
GetMatrixBfromCoefs(coefs, freqs, dt = 0.25, a0 = NULL)
}
\arguments{
\item{coefs}{Coefficients from a var model}

\item{freqs}{A vector of frequencies}

\item{dt}{Inverse sample rate. Default is 0.25}

\item{a0}{Use an a0 function (only valid if the coefficients have been
previously adjusted with an a0 function). The default is NULL,
in which case, the identity matrix I is assumed to be the a0 function.}
}
\value{
The auxiliary matrix B
}
\description{
Calculates transfer function B from a specific VAR model coefficients
}
\references{
Barbieri R, Parati G, Saul JP. Closed- versus Open-Loop Assessment of Heart Rate
Baroreflex. IEEE Eng Med Biol Mag. 2001;20(2):33-42

Korhonen I, Mainardi L, Baselli G, Bianchi A, Loula P, 
Carrault. Linear multivariate models for physiological signal analysis: applications. 
Comput Methods Programs Biomed. 1996;51(1-2):121-30

Faes L, Nollo G. Multivariate Frequency Domain Analysis of Causal Interactions
in Physiological Time Series. In Biomedical Engineering, Trends in Electronics,
Communications and Software. 2011

Hytti H, Takalo R, Ihalainen H. Tutorial on Multivariate Autoregressive Modelling.
J Clin Monit Comput. 2006;20(1):101-8
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
\keyword{internal}
