% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_Modelling.R
\name{GetMatrixBfromVAR}
\alias{GetMatrixBfromVAR}
\title{Calculate transfer function B from VAR model}
\usage{
GetMatrixBfromVAR(var, freqs, dt = 0.25, a0 = FALSE)
}
\arguments{
\item{var}{A var model}

\item{freqs}{A vector of frequencies}

\item{dt}{Inverse sample rate. Default is 0.25}

\item{a0}{Boolean. Should the coefficients of the VAR model be adjusted with 
instantaneous interactions? Default is FALSE}
}
\value{
The auxiliary matrix B
}
\description{
Calculates transfer function B from a specific VAR model
}
\references{
Barbieri R, Parati G, Saul JP. Closed- versus Open-Loop Assessment of Heart Rate
Baroreflex. IEEE Eng Med Biol Mag. 2001;20(2):33-42

Korhonen I, Mainardi L, Baselli G, Bianchi A, Loula P, 
Carrault. Linear multivariate models for physiological signal analysis: applications. 
Comput Methods Programs Biomed. 1996;51(1-2):121-30


Faes L, Nollo G. Multivariate Frequency Domain Analysis of Causal Interactions
in Physiological Time Series. In Biomedical Engineering, Trends in Electronics,
Communications and Software. 2011

Hytti H, Takalo R, Ihalainen H. Tutorial on Multivariate Autoregressive Modelling.
J Clin Monit Comput. 2006;20(1):101-8
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
\keyword{internal}
