% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Causal Coherence.R
\name{GetMaxCoherence}
\alias{GetMaxCoherence}
\title{Get maximum coherence}
\usage{
GetMaxCoherence(SM, coherence, HF = 0.4, LF = 0.15, VLF = 0.04)
}
\arguments{
\item{SM}{the computed frequency domain model from the system to be analyzed}

\item{coherence}{the coherence function to be decomposed}

\item{HF}{maximum limit for the HF band. Default is 0.4}

\item{LF}{maximum limit for the LF band. Default is 0.15}

\item{VLF}{maximum limit for the VLF band. Default is 0.04}
}
\value{
A list with the computed maximum values at the LF and HF bands for the 
        coherence
}
\description{
Computes the maximum coherence values for the LF and HF bands
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)
coh <- abs(CalculateCoherence(freq_model))^2
GetMaxCoherence(freq_model, coh)
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
