% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phases.R
\name{GetOpenTransFunPhase}
\alias{GetOpenTransFunPhase}
\title{Calculate open-loop phase transfer function}
\usage{
GetOpenTransFunPhase(SM, index1, index2)
}
\arguments{
\item{SM}{the computed frequency domain model from the system to be analyzed}

\item{index1}{numeric index to specify a variable from the model}

\item{index2}{numeric index to specify a variable from the model}
}
\value{
A vector containing the phase transfer function obtained from the analyzed
        open-loop transfer function
}
\description{
Estimates the phase transfer function out of the open-loop transfer functions
of a particular system
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)
Phase_open <- GetOpenTransFunPhase(model, 1, 2)
Phase_open[1:100]
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
