% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolate Data.R
\name{InterpolateData}
\alias{InterpolateData}
\title{Interpolate data}
\usage{
InterpolateData(x, f = 4)
}
\arguments{
\item{x}{a list containing data to be interpolated}

\item{f}{sample rate for the interpolated series. Default is 4 Hz}
}
\value{
A list containing the interpolated data.
}
\description{
Interpolates data up to a particular sample rate
}
\examples{
data(Cardiovascular)
interpolated <- InterpolateData(Cardiovascular)
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
