% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param Modelling.R
\name{ParamFreqModel}
\alias{ParamFreqModel}
\title{Parametric frequency model estimation}
\usage{
ParamFreqModel(
  model,
  len = 1000,
  dt = 0.25,
  A0 = TRUE,
  sigma = NULL,
  coefs = NULL
)
}
\arguments{
\item{model}{a time domain VAR model}

\item{len}{length of the vector of frequencies. Default is 1000}

\item{dt}{inverse of sample rate. Default is 0.25}

\item{A0}{boolean; add 0 effects. Default TRUE}

\item{sigma}{include a specific covariance matrix; otherwise it is obtained
from the VAR model. Default is NULL}

\item{coefs}{include VAR coefficients; otherwise these are obtained from the
VAR model. Default is NULL}
}
\value{
A list with the estimated components of the frequency domain model.
        These components include closed-loop transfer functions, open-loop
        transfer functions, open-vs-closed-loop overestimations, noise transfer
        functions, and spectral estimations of the variables.
}
\description{
Estimates frequency model parameters from a time domain VAR model.
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)
}
\references{
Barbieri R, Parati G, Saul JP. Closed- versus Open-Loop Assessment of Heart Rate
Baroreflex. IEEE Eng Med Biol Mag. 2001;20(2):33-42

Korhonen I, Mainardi L, Baselli G, Bianchi A, Loula P, 
Carrault. Linear multivariate models for physiological signal analysis: applications. 
Comput Methods Programs Biomed. 1996;51(1-2):121-30


Faes L, Nollo G. Multivariate Frequency Domain Analysis of Causal Interactions
in Physiological Time Series. In Biomedical Engineering, Trends in Electronics,
Communications and Software. 2011

Hytti H, Takalo R, Ihalainen H. Tutorial on Multivariate Autoregressive Modelling.
J Clin Monit Comput. 2006;20(1):101-8
}
