% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Causal Coherence.R
\name{PlotCausalCoherence}
\alias{PlotCausalCoherence}
\title{Plot Causal coherence}
\usage{
PlotCausalCoherence(SM, CCoh, VLF = 0.04, LF = 0.15, HF = 0.4, xlim = NULL)
}
\arguments{
\item{SM}{the computed frequency domain model from the system to be analyzed}

\item{CCoh}{the computed causal coherence from the system to be analyzed}

\item{VLF}{maximum limit for the VLF band. Default is 0.04}

\item{LF}{maximum limit for the LF band. Default is 0.15}

\item{HF}{maximum limit for the HF band. Default is 0.4}

\item{xlim}{vector with minimum and maximum limits for the graphical representation.
Default is NULL}
}
\value{
None
}
\description{
Plot Causal coherence
}
\examples{
#ADD EXAMPLE
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
