% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noise_Contribution.R
\name{PlotCausality}
\alias{PlotCausality}
\title{Plot causal flow}
\usage{
PlotCausality(SM, index, VLF = 0.04, LF = 0.15, HF = 0.4, xlim = NULL)
}
\arguments{
\item{SM}{the computed frequency domain model from the system to be analyzed}

\item{index}{an numeric value indicating which branch should be highlited}

\item{VLF}{maximum limit for the VLF band. Default is 0.04}

\item{LF}{maximum limit for the LF band. Default is 0.15}

\item{HF}{maximum limit for the HF band. Default is 0.4}

\item{xlim}{specific limits for the frequency axis. Default is NULL}
}
\value{
None
}
\description{
Generates frequency domain plots indicating the behavior of the causal flow of a certain branch
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)
PlotCausality(freq_model, 1)
PlotCausality(freq_model, 2)
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
