% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coherence.R
\name{PlotCoherence}
\alias{PlotCoherence}
\title{Plot spectral coherence}
\usage{
PlotCoherence(
  SM,
  index1,
  index2,
  coherence,
  thr = 0.5,
  xlim = NULL,
  HFcolor = "yellow",
  LFcolor = "green",
  VLF = 0.04,
  LF = 0.15,
  HF = 0.4,
  show.cols = TRUE,
  thr.col = "red",
  phase.col = "red"
)
}
\arguments{
\item{SM}{the computed frequency domain model from the system to be analyzed}

\item{index1}{numeric index to specify a variable from the model}

\item{index2}{numeric index to specify a variable from the model}

\item{coherence}{a vector with the spectral coherence values}

\item{thr}{a specific coherence threshold. Default is 0.5}

\item{xlim}{specific limits for the frequency axis. Default is NULL}

\item{HFcolor}{color for the HF band. Default is yellow}

\item{LFcolor}{color for the LF band. Default is green}

\item{VLF}{maximum limit for the VLF band. Default is 0.04}

\item{LF}{maximum limit for the LF band. Default is 0.15}

\item{HF}{maximum limit for the HF band. Default is 0.4}

\item{show.cols}{boolean, show colors in the plot. Default is TRUE}

\item{thr.col}{color for the coherence threshold. Default is red}

\item{phase.col}{color for the phase difference plot. Default is red}
}
\value{
None
}
\description{
Plots the spectral coherence between two variables
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)

coherence <- CalculateCoherence(freq_model, 1, 2)

PlotCoherence(freq_model, 1, 2, coherence)

}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
