% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{PlotRoots}
\alias{PlotRoots}
\title{Plot VAR model inverse roots}
\usage{
PlotRoots(model, col = "blue", col2 = "red")
}
\arguments{
\item{model}{a VAR model to be checked}

\item{col}{color for the roots inside the unit circle. Default is blue}

\item{col2}{color for the roots outside the unit circle. Default is red}
}
\value{
A boolean variable indicating if the multivariate time series is stationary or not,
        or a caption indicating so, whether the argument verbose is TRUE or
        FALSE
}
\description{
Plot a unit circle with inverse roots computed from a VAR model
}
\examples{
data(DetrendedData)

model <- EstimateVAR(DetrendedData)
PlotRoots(model)
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
\keyword{internal}
