% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate from model.R
\name{PlotSimulatedS}
\alias{PlotSimulatedS}
\title{Plot simulated model}
\usage{
PlotSimulatedS(
  SM,
  S,
  index,
  unit = "ms2/Hz",
  xlim = NULL,
  HFcolor = "yellow",
  LFcolor = "green",
  VLF = 0.04,
  LF = 0.15,
  HF = 0.4,
  show.cols = TRUE,
  phase.col = "red"
)
}
\arguments{
\item{SM}{the computed frequency domain model from the system to be analyzed}

\item{S}{simulated spectral matrix from the model to be analyzed}

\item{index}{numeric index to specify a variable from the model}

\item{unit}{unit of the output variable. Default is ms2/Hz}

\item{xlim}{specific limits for the frequency axis. Default is NULL}

\item{HFcolor}{color for the HF band. Default is yellow}

\item{LFcolor}{color for the LF band. Default is green}

\item{VLF}{maximum limit for the VLF band. Default is 0.04}

\item{LF}{maximum limit for the LF band. Default is 0.15}

\item{HF}{maximum limit for the HF band. Default is 0.4}

\item{show.cols}{boolean, show colors in the plot. Default is TRUE}

\item{phase.col}{color for the phase difference plot. Default is red}
}
\value{
None
}
\description{
Generates a graphical representation of the noise transfer functions
}
\examples{
# ADD EXAMPLE

}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
