% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotTransferFun}
\alias{PlotTransferFun}
\title{Plot transfer functions}
\usage{
PlotTransferFun(
  SM,
  index1,
  index2,
  unit1 = "ms",
  unit2 = "mmHg",
  xlim = NULL,
  HFcolor = "yellow",
  LFcolor = "green",
  VLF = 0.04,
  LF = 0.15,
  HF = 0.4,
  show.cols = TRUE,
  phase.col = "red",
  open = FALSE,
  plot.phase = TRUE
)
}
\arguments{
\item{SM}{the computed frequency domain model from the system to be analyzed}

\item{index1}{numeric index to specify a variable from the model}

\item{index2}{numeric index to specify a variable from the model}

\item{unit1}{unit of the output variable. Default is ms}

\item{unit2}{unit of the input variable. Default is mmHg}

\item{xlim}{specific limits for the frequency axis. Default is NULL}

\item{HFcolor}{color for the HF band. Default is yellow}

\item{LFcolor}{color for the LF band. Default is green}

\item{VLF}{maximum limit for the VLF band. Default is 0.04}

\item{LF}{maximum limit for the LF band. Default is 0.15}

\item{HF}{maximum limit for the HF band. Default is 0.4}

\item{show.cols}{boolean, show colors in the plot. Default is TRUE}

\item{phase.col}{color for the phase difference plot. Default is red}

\item{open}{boolean, show open-loop transfer functions. Default is FALSE}

\item{plot.phase}{boolean, plot phase difference calculated from the transfer function.
Default is TRUE}
}
\value{
None
}
\description{
Generates a graphical representation of the closed-loop and open-loop transfer
functions
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)

coherence <- CalculateCoherence(freq_model, 1, 2)

PlotCoherence(freq_model, 1, 2, coherence)

}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
