% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate from model.R
\name{PSD}
\alias{PSD}
\title{Compute and plot spectral density}
\usage{
PSD(
  x,
  p,
  xlim = NULL,
  HFcolor = "yellow",
  LFcolor = "green",
  VLF = 0.04,
  LF = 0.15,
  HF = 0.4,
  show.cols = TRUE,
  phase.col = "red",
  f = 4,
  plot = TRUE,
  output = FALSE,
  n = 1500,
  unit = "ms2/Hz"
)
}
\arguments{
\item{x}{a vector with observations to be analyzed}

\item{p}{model order for the autoregressive model}

\item{xlim}{specific limits for the frequency axis. Default is NULL}

\item{HFcolor}{color for the HF band. Default is yellow}

\item{LFcolor}{color for the LF band. Default is green}

\item{VLF}{maximum limit for the VLF band. Default is 0.04}

\item{LF}{maximum limit for the LF band. Default is 0.15}

\item{HF}{maximum limit for the HF band. Default is 0.4}

\item{show.cols}{boolean, show colors in the plot. Default is TRUE}

\item{phase.col}{color for the phase difference plot. Default is red}

\item{f}{sample rate of the vector of observations. Default is 4 Hz}

\item{plot}{boolean, plot the PSD function. Default is TRUE}

\item{output}{boolean, return the PSD function. Default is FALSE}

\item{n}{length of the vector of frequencies. Default is 1500}

\item{unit}{unit of the output variable. Default is ms2/Hz}
}
\value{
If the output argument is TRUE, a list with the computed power at
        LF and HF bands, as well as the peak frequencies, is returned.
}
\description{
Computes and plots spectral density estimates from a certain variable
}
\examples{
data(DetrendedData)
RR <- DetrendedData[,"RR"]

PSD(RR, 21)

HRV <- PSD(RR, 21, plot = FALSE, output = TRUE)

}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
