% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resample_Data.R
\name{ResampleData}
\alias{ResampleData}
\title{Resample data}
\usage{
ResampleData(x, f = 4)
}
\arguments{
\item{x}{a list containing data to be interpolated}

\item{f}{sample rate for the interpolated series. Default is 4 Hz}
}
\value{
A list containing the interpolated data.
}
\description{
Resamples data up to a particular sample rate
}
\examples{
data(Cardiovascular)
interpolated <- ResampleData(Cardiovascular)
plot(interpolated$Time, interpolated$RR, xlab = "Time (sec)", ylab = "RR intervals (ms)")
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
