% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_from_model.R
\name{SimulateWithModel}
\alias{SimulateWithModel}
\title{Simulate frequency domain model}
\usage{
SimulateWithModel(
  SM,
  noises,
  a0,
  f,
  xlim = NULL,
  HFcolor = "yellow",
  LFcolor = "green",
  VLF = 0.04,
  LF = 0.15,
  HF = 0.4,
  show.cols = TRUE,
  phase.col = "red"
)
}
\arguments{
\item{SM}{A computed frequency domain model}

\item{noises}{Which noise spectra to be included in the simulation}

\item{a0}{Which immediate effects to be included in the simulation}

\item{f}{Sample frequency}

\item{xlim}{Specific limits for the frequency axis. Default is NULL}

\item{HFcolor}{Color for the HF band. Default is yellow}

\item{LFcolor}{color for the LF band. Default is green}

\item{VLF}{maximum limit for the VLF band. Default is 0.04}

\item{LF}{maximum limit for the LF band. Default is 0.15}

\item{HF}{maximum limit for the HF band. Default is 0.4}

\item{show.cols}{boolean, show colors in the plot. Default is TRUE}

\item{phase.col}{color for the phase difference plot. Default is red}
}
\value{
A modified version of the original model according to the chosen parameters
}
\description{
Simulates the behavior of a frequency domain model after choosing several variables
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)
new_model <- SimulateWithModel(freq_model, c(2,3), a0 = 2, f = 4)
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
