% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param Modelling.R
\name{UpdateWithA0}
\alias{UpdateWithA0}
\title{Update coefficients with A0 function}
\usage{
UpdateWithA0(A0, coefs)
}
\arguments{
\item{A0}{The A0 function}

\item{coefs}{Coefficients to be updated}
}
\value{
The updated coefficients
}
\description{
Updates VAR model coefficients using an A0 function
}
\examples{
#ADD EXAMPLE
}
\references{
Barbieri R, Parati G, Saul JP. Closed- versus Open-Loop Assessment of Heart Rate
Baroreflex. IEEE Eng Med Biol Mag. 2001;20(2):33-42

Korhonen I, Mainardi L, Baselli G, Bianchi A, Loula P, 
Carrault. Linear multivariate models for physiological signal analysis: applications. 
Comput Methods Programs Biomed. 1996;51(1-2):121-30


Faes L, Nollo G. Multivariate Frequency Domain Analysis of Causal Interactions
in Physiological Time Series. In Biomedical Engineering, Trends in Electronics,
Communications and Software. 2011

Hytti H, Takalo R, Ihalainen H. Tutorial on Multivariate Autoregressive Modelling.
J Clin Monit Comput. 2006;20(1):101-8
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
\keyword{internal}
