% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Causal Coherence.R
\name{CalculateCausalCoherence}
\alias{CalculateCausalCoherence}
\title{Calculate causal coherence from a system modeled in the frequency domain}
\usage{
CalculateCausalCoherence(SM, Mod = TRUE)
}
\arguments{
\item{SM}{The computed frequency domain model from the system to be analyzed}

\item{Mod}{Compute absolute values. Default is TRUE}
}
\value{
A list containing the following components: 
\item{C1}{Causal coherence from y to x}
\item{C2}{Causal coherence from x to y} 
\item{C3}{The standard, non-causal squared coherence}
}
\description{
Calculate causal coherence from a system modeled in the frequency domain
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)
ccoh <- CalculateCausalCoherence(freq_model)
}
\references{
Faes L, Porta A, Antolini R, Nollo G. Role of causality in the evaluation of coherence and 
transfer function between heart period and systolic pressure in humans. Comput Cardiol.
2004;277-80.

Faes L, Mase M, Nollo G, Chon KH, Florian JP. Measuring postural-related changes of spontaneous
baroreflex sensitivity after repeated long-duration diving: frequency domain approaches. Auton Neurosci. 
2013 Nov;178(1-2):96-102.
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
