% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_Stats.R
\name{CheckStationarity}
\alias{CheckStationarity}
\title{Check stationarity}
\usage{
CheckStationarity(
  x,
  alpha = 0.05,
  lags = NULL,
  warnings = TRUE,
  verbose = FALSE,
  correction = "bonferroni",
  verbose.method = c("c", "p")
)
}
\arguments{
\item{x}{a matrix with the data to be checked. Number of columns should be equal
to the number of variables.}

\item{alpha}{significance level for testing. Default is 0.05}

\item{lags}{specify a certain level of lags for the test. Default is NULL}

\item{warnings}{boolean, show warnings regarding the test results. Default is TRUE}

\item{verbose}{boolean, show results from the test in captions. Default is FALSE}

\item{correction}{choose a p-value correction method for multiple hypotheses. Default
is bonferroni. For other methods, please check \link[stats]{p.adjust}
and \link[stats]{p.adjust.methods} from package
\href{https://CRAN.R-project.org/package=stats}{stats}}

\item{verbose.method}{method for the verbose option, either "c" (cat) or
"p" (print). Default is c.}
}
\value{
A boolean variable indicating if the multivariate time series is stationary or not,
        or a caption indicating so, depending on whether the arguments verbose and warning are TRUE or
        FALSE
}
\description{
Check stationarity of the data
}
\details{
Several functions are available in this package to statistically validate
properties of the evaluated time series and VAR models. These functions are based
on the validation algorithms proposed in Seth (2010) and Seth (2011) for MVAR validation.
The functions allow corrections for multiple comparisons offered by \link[stats]{p.adjust} 
and in \link[stats]{p.adjust.methods}. The bonferroni correction has been selected as
default, as suggested in Seth (2010). 
 
In this function, stationarity of the time series is evaluated according to the criteria
established in Seth (2010) and more specifically in Seth (2011).
}
\examples{
data(DetrendedData)

CheckStationarity(DetrendedData)
CheckStationarity(DetrendedData, verbose = TRUE)
}
\references{
Seth AK. A MATLAB toolbox for Granger causal connectivity analysis. J 
Neurosci Methods. 2010;186(2):262-73.

Seth AK. Granger Causal Conectivity Analysis: A MATLAB Toolbox. 2011.

Ding M, Bressler S, Yang W, Liang H. Short-window spectral analysis of cortical 
event-related potentials by adaptative multivariate autoregressive modelling: data
preprocessing, model validation, and variability assessment. Biol Cybern. 2000;83:35-45
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
