% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detrend.R
\name{DetrendByCutoff}
\alias{DetrendByCutoff}
\title{Detrend variables}
\usage{
DetrendByCutoff(x, cutoff = 0.04, f = 4, wv = "d16", max_f = 0.4)
}
\arguments{
\item{x}{a vector of observations}

\item{cutoff}{a frequency used as limit for the detrending. Default is 0.04 Hz}

\item{f}{sample rate of the vector of observations. Default is 4 Hz}

\item{wv}{wavelet to be passed into the \link[waveslim]{modwt} function. Default is
d16}

\item{max_f}{maximum frequency to be accounted for in the detrending. default is 0.4 Hz}
}
\value{
The detrended vector of observations
}
\description{
Detrends a vector of observations using the discrete wavelet transform
}
\examples{
data(Cardiovascular)
data <- InterpolateData(data)
RR <- DetrendByCutoff(data$RR) # The RR series is detrended
SBP <- DetrendByCutoff(data$SBP) # The SBP series is detrended
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
