% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{DiagnoseResiduals}
\alias{DiagnoseResiduals}
\title{Check white-noise nature of residuals}
\usage{
DiagnoseResiduals(
  model,
  alpha = 0.05,
  correction = "bonferroni",
  verbose = FALSE
)
}
\arguments{
\item{model}{a VAR model to be checked}

\item{alpha}{significance level for testing. Default is 0.05}

\item{correction}{choose a p-value correction method for multiple hypotheses. Default
is bonferroni. For other methods, please check \link[stats]{p.adjust.methods}}

\item{verbose}{boolean, show results from the test in captions. Default is FALSE}

\item{lags}{specify a certain level of lags for the test. Default is NULL}
}
\value{
A boolean variable indicating if the multivariate time series is stationary or not,
        or a caption indicating so, whether the argument verbose is TRUE or
        FALSE
}
\description{
Check that residuals from a VAR model are a white-noise process
}
\examples{
data(DetrendedData)

model <- EstimateVAR(DetrendedData)
DiagnoseResiduals(model)
}
\references{
Seth AK. A MATLAB toolbox for Granger causal connectivity analysis. J 
Neurosci Methods. 2010;186(2):262-73.

Seth AK. Granger Causal Conectivity Analysis: A MATLAB Toolbox. 2011.

Ding M, Bressler S, Yang W, Liang H. Short-window spectral analysis of cortical 
event-related potentials by adaptative multivariate autoregressive modelling: data
preprocessing, model validation, and variability assessment. Biol Cybern. 2000;83:35-45
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
