% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{DiagnoseStability}
\alias{DiagnoseStability}
\title{Check stability of a VAR model}
\usage{
DiagnoseStability(
  model,
  do.plot = FALSE,
  verbose = FALSE,
  col = "blue",
  col2 = "red"
)
}
\arguments{
\item{model}{a VAR model to be checked}

\item{do.plot}{boolean, plot a unitary circle showing the inverse roots of the
model. Default is FALSE}

\item{verbose}{boolean, show results from the test in captions. Default is FALSE}

\item{col}{color for the roots inside the unit circle. Default is blue}

\item{col2}{color for the roots outside the unit circle. Default is red}
}
\value{
A boolean variable indicating if the multivariate time series is stationary or not,
        or a caption indicating so, whether the argument verbose is TRUE or
        FALSE
}
\description{
Check that the estimated VAR model is stable
}
\examples{
data(DetrendedData)

model <- EstimateVAR(DetrendedData)
DiagnoseStability(model, do.plot = TRUE)
}
\references{
Seth AK. A MATLAB toolbox for Granger causal connectivity analysis. J 
Neurosci Methods. 2010;186(2):262-73.

Seth AK. Granger Causal Conectivity Analysis: A MATLAB Toolbox. 2011.

Ding M, Bressler S, Yang W, Liang H. Short-window spectral analysis of cortical 
event-related potentials by adaptative multivariate autoregressive modelling: data
preprocessing, model validation, and variability assessment. Biol Cybern. 2000;83:35-45
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
