% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{EstimateVAR}
\alias{EstimateVAR}
\title{Estimate a VAR model}
\usage{
EstimateVAR(
  x,
  pmax = 22,
  p = NULL,
  ic = "AIC",
  alpha = 0.05,
  correction = "bonferroni",
  type = "none",
  exogen = NULL,
  warnings = TRUE
)
}
\arguments{
\item{x}{a matrix with the data to be checked. Number of columns should be equal
to the number of variables.}

\item{pmax}{the maximum order of the model to be estimated. Default is 22.}

\item{p}{a specific model order for the VAR model. Default is NULL}

\item{ic}{an information criterium to estimate the model order, if not previously
specified. Default is AIC, the Akaike criterium}

\item{alpha}{significance level for testing. Default is 0.05}

\item{correction}{choose a p-value correction method for multiple hypotheses. Default
is bonferroni. For other methods, please check \link[stats]{p.adjust.methods}}

\item{type}{parameter to be passed into the \link[vars]{VAR} function, indicating the type of model.
Default is none}

\item{exogen}{parameter to be passed into the \link[vars]{VAR} function, indicating 
exogenous variables. Default is NULL}

\item{warnings}{boolean, show warnings regarding the test results. Default is FALSE}
}
\value{
A boolean variable indicating if the multivariate time series is stationary or not,
        or a caption indicating so, depending on whether the arguments verbose and warning are TRUE or
        FALSE
}
\description{
Estimates and validates a VAR model
}
\details{
This function estimates a particular VAR model, and at the same
time, checks the different properties of the variables and the model and thus informs about
the model's validity. For mor information regarding the VAR model estimation, please check the 
references section.
}
\examples{
data(DetrendedData)

CheckStationarity(DetrendedData)
CheckStationarity(DetrendedData, verbose = TRUE)
}
\references{
Bernhard Pfaff (2008). VAR, SVAR and SVEC Models: Implementation Within R Package vars. Journal of
  Statistical Software 27(4). URL https://www.jstatsoft.org/v27/i04/.

  Pfaff, B. (2008) Analysis of Integrated and Cointegrated Time Series with R. Second Edition. Springer,
  New York. ISBN 0-387-27960-1
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
