% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expected_Values_Freq_Domain.R
\name{GetPeaks}
\alias{GetPeaks}
\title{Get peaks for each band}
\usage{
GetPeaks(SM, VLF = 0.04, LF = 0.15, HF = 0.4, str = TRUE)
}
\arguments{
\item{SM}{the computed frequency domain model from the system to be analyzed}

\item{VLF}{maximum limit for the VLF band. Default is 0.04}

\item{LF}{maximum limit for the LF band. Default is 0.15}

\item{HF}{maximum limit for the HF band. Default is 0.4}

\item{str}{boolean, if TRUE the output will be the structure of the returned R object.
Default is FALSE}
}
\value{
A list containing the estimates at the HF and LF bands; or
        the structure of this list.
}
\description{
Get the frequency values in which the maximum peaks are located for each band
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)
GetPeaks(freq_model)
Peaks <- GetPeaks(freq_model, str = FALSE)
Peaks

}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
