% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param Modelling.R
\name{IncludeA0Effects}
\alias{IncludeA0Effects}
\title{Include A0 effects}
\usage{
IncludeA0Effects(system)
}
\arguments{
\item{system}{the model of the system whose instantaneous effects we want to include}
}
\value{
A list with the estimated components of the frequency domain model.
        These components include closed-loop transfer functions, open-loop
        transfer functions, open-vs-closed-loop overestimations, noise transfer
        functions, and spectral estimations of the variables.
}
\description{
Estimates and includes A0 effects for a frequency domain model
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model, A0 = FALSE)

#To include the A0 function:
freq_model <- IncludeA0Effects(freq_model)
}
\references{
Barbieri R, Parati G, Saul JP. Closed- versus Open-Loop Assessment of Heart Rate
Baroreflex. IEEE Eng Med Biol Mag. 2001;20(2):33-42

Korhonen I, Mainardi L, Baselli G, Bianchi A, Loula P, 
Carrault. Linear multivariate models for physiological signal analysis: applications. 
Comput Methods Programs Biomed. 1996;51(1-2):121-30


Faes L, Nollo G. Multivariate Frequency Domain Analysis of Causal Interactions
in Physiological Time Series. In Biomedical Engineering, Trends in Electronics,
Communications and Software. 2011

Hytti H, Takalo R, Ihalainen H. Tutorial on Multivariate Autoregressive Modelling.
J Clin Monit Comput. 2006;20(1):101-8
}
