% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametric_Modelling.R
\name{ParamFreqModel}
\alias{ParamFreqModel}
\title{Parametric frequency model estimation}
\usage{
ParamFreqModel(
  model,
  len = 1000,
  dt = 0.25,
  A0 = TRUE,
  sigma = NULL,
  coefs = NULL,
  freq_support = c("pgram", "seq"),
  use.cross = TRUE,
  preserve.signals = TRUE,
  preserve.p = TRUE
)
}
\arguments{
\item{model}{A time domain VAR model}

\item{len}{Length of the vector of frequencies. Default is 1000}

\item{dt}{Inverse of sample rate. Default is 0.25}

\item{A0}{Boolean; add 0 effects. Default TRUE}

\item{sigma}{Include a specific covariance matrix; otherwise it is obtained
from the VAR model. Default is NULL}

\item{coefs}{Include VAR coefficients; otherwise these are obtained from the
VAR model. Default is NULL}

\item{freq_support}{Method to compute the frequency support vector, either "pgram" or "seq".
Default is "pgram".}

\item{use.cross}{Boolean; use cross-spectrum to compute the open transfer functions.
default is TRUE. If FALSE, an alpha-index is used to compute the
open transfer functions}

\item{preserve.signals}{Boolean, include the input and output signals into the model list. Default
is TRUE}

\item{preserve.p}{Boolean, include VAR model order p into the model list. Default
is TRUE}
}
\value{
A list with the estimated components of the frequency domain model:
\item{Freqs}{Vector of frequency values}
\item{Vars_Transfer_funs}{Frequency-domain represantations of the VAR model blocks} 
\item{Transfer_Functions}{The computed closed-loop baroreflex transfer functions}
\item{Open_Transfer_Functions}{The computed open-loop baroreflex transfer functions}
\item{Open_Transfer_Functions_Overestimation}{How the open loop condition overestimates the results}
\item{Noise_Transfer_Functions}{The computed transfer functions from the noise source model}
\item{Spectra}{The computed auto-spectra and cross-spectra of the variables}
\item{Noise_Spectra}{The spectra of the white noise sources}
\item{a0}{The computed immediate effects of the model}
}
\description{
Estimates frequency model parameters from a time domain VAR model.
}
\details{
This function calculates the frequency domain components and transfer functions
from the time domain MVAR model. The functions are estimated taking into account the following representations 
of the model:

\eqn{X = BX + W}
\eqn{AX = W}
\eqn{X = HW}

So that:

\eqn{S = H * sigma * t(Conj(H))}
\eqn{A * S * t(Conj(A)) = sigma}

By allowing immediate effects from SBP to RR, the baroreflex transfer function
is calculated as:

\eqn{b_{21}/(1-b_{22}) = -a_{21}/a_{22} = h_{21}/h_{11}}


Some of the modules integrated in these functions have been designed so that the outputs
have compatible characteristics (and are thus reproducible) with the ones obtained from 
other packages such as \href{https://CRAN.R-project.org/package=grangers}{package grangers}
(e.g., the freq_support option available in this function allows said compatibility). For
more information, please check the references section.
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)
}
\references{
Barbieri R, Parati G, Saul JP. Closed- versus Open-Loop Assessment of Heart Rate
Baroreflex. IEEE Eng Med Biol Mag. 2001;20(2):33-42

Korhonen I, Mainardi L, Baselli G, Bianchi A, Loula P, 
Carrault. Linear multivariate models for physiological signal analysis: applications. 
Comput Methods Programs Biomed. 1996;51(1-2):121-30

Faes L, Nollo G. Multivariate Frequency Domain Analysis of Causal Interactions
in Physiological Time Series. In Biomedical Engineering, Trends in Electronics,
Communications and Software. 2011

Hytti H, Takalo R, Ihalainen H. Tutorial on Multivariate Autoregressive Modelling.
J Clin Monit Comput. 2006;20(1):101-8

Matteo Farne and Angela Montanari (2019). grangers: Inference on Granger-Causality 
in the Frequency Domain. R package version 0.1.0. https://CRAN.R-project.org/package=grangers
}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
