% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noise Contribution.R
\name{PlotNoiseContribution}
\alias{PlotNoiseContribution}
\title{Plot noise source constribution}
\usage{
PlotNoiseContribution(cont1, cont2, label1 = "RR noise", label2 = "SBP noise")
}
\arguments{
\item{cont1}{estimated contribution from one noise source}

\item{cont2}{estimated contribution from a second noise source}

\item{label1}{label for the first noise source. Default is "RR noise"}

\item{label2}{label for the second noise source. Default is "RR noise"}
}
\value{
None
}
\description{
Generates pie plots showing the estimated noise source contribution from two noise sources
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)

noise_con1 <- NoiseSourceContribution(freq_model, 1, 2, use.coh = FALSE)
noise_con2 <- NoiseSourceContribution(freq_model, 1, 1, use.coh = FALSE)
noise_con1
noise_con2
noise_con1 + noise_con2
PlotNoiseContribution(noise_con1, noise_con2)

}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
