% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_from_model.R
\name{PlotSimulatedS}
\alias{PlotSimulatedS}
\title{Plot simulated model}
\usage{
PlotSimulatedS(
  SM,
  S,
  index,
  unit = "ms2/Hz",
  xlim = NULL,
  HFcolor = "yellow",
  LFcolor = "green",
  VLF = 0.04,
  LF = 0.15,
  HF = 0.4,
  show.cols = TRUE,
  phase.col = "red"
)
}
\arguments{
\item{SM}{The computed frequency domain model from the system to be analyzed}

\item{S}{Simulated spectral matrix from the model to be analyzed}

\item{index}{Numeric index to specify a variable from the model}

\item{unit}{Unit of the output variable. Default is ms2/Hz}

\item{xlim}{Specific limits for the frequency axis. Default is NULL}

\item{HFcolor}{Color for the HF band. Default is yellow}

\item{LFcolor}{Color for the LF band. Default is green}

\item{VLF}{Maximum limit for the VLF band. Default is 0.04}

\item{LF}{Maximum limit for the LF band. Default is 0.15}

\item{HF}{Maximum limit for the HF band. Default is 0.4}

\item{show.cols}{Boolean, show colors in the plot. Default is TRUE}

\item{phase.col}{Color for the phase difference plot. Default is red}
}
\value{
None
}
\description{
Generates a graphical representation of the noise transfer functions
}
\examples{
data(DetrendedData)
model <- EstimateVAR(DetrendedData)
freq_model <- ParamFreqModel(model)
new_model <- SimulateWithModel(freq_model, c(2,3), a0 = 2, f = 4)
PlotSimulatedS(freq_model, new_model, 2)

}
\author{
Alvaro Chao-Ecija, Marc Stefan Dawid-Milner
}
